change buffer size for m1s_xram_audio_init
I might be wrong on this, but when I used the example code as-is, I would get memory corruption.  Though I haven't found any documentation on the SDK, I suspect that the buffer size parameter of m1s_xram_audio_init should not be AUDIO_BUFF_SIZE*2.  When I changed it to AUDIO_BUFF_SIZE, the memory corruption went away.